requirements__.txt
django
khayyam
pillow
django-tinymce
djangorestframework
openpyxl